Commented and changed files
I had all the text and the buttons to be aligned in the center. I also added the values on the screen for the buttons as I noticed you did not have any. Made the text black as the text was hard to see in the sky background. I commented on your javascript and somewhat made it organized. I gave you an introduction to create a reset button later if you want to implement cookies.